fix database problemss with user and role

+++ src/main/resources/db/migration/V7__enhance_problems_and_add_tags.sql
@@ -9,8 +9,8 @@ CREATE TABLE tags (
 
 -- Add problem_tags junction table (many-to-many)
 CREATE TABLE problem_tags (
-    problem_id INTEGER NOT NULL,
-    tag_id INTEGER NOT NULL,
+    problem_id BIGINT NOT NULL,
+    tag_id BIGINT NOT NULL,
     PRIMARY KEY (problem_id, tag_id),
     FOREIGN KEY (problem_id) REFERENCES problems(id) ON DELETE CASCADE,
     FOREIGN KEY (tag_id) REFERENCES tags(id) ON DELETE CASCADE
